Create folders for each main section
It will be much easier in the future to have all markdown files witin different folders (a folder represnts the parent node in a tree structure). Like this, we will have the same strutucre as UI and as folders.

Let's supose that a user want to create a new subsection within the Learn Cardano section, then he will need to create a folder that has the name of that subsection and indside it include the article.