Add explorer link and made them both English (#276)
There is a link to the explorer above, but not actually in the step where you need to access the explorer. Added a link to make it a bit easier for users to follow along in case they missed the link previously. Also, I switched both links to the English landing page vs German since the tutorial is in English.