Notes on versions of compiler, random and wallet
I've added notes to references of the compiler, random, and wallet versions used in this article as they are outdated. Ideally, the code would be changed so that these instructions work regardless of the version of the code that is current. These changes are beyond my current level of involvement with the project