Use stack install command to install cardano-wallet
This commit update the documentation section for the cardano-wallet.
As the build instruction is based on the `stack` tool then it makes
sense use the `stack install` sub command to install produced binaries.

`stack install` copies binaries into $HOME/.local/bin/.
As of now copied binaries are: cardano-wallet, local-cluser and
mock-token-metadata-server.

By effect this commit removes the need to have a macos / Linux tabs
and thus simplifies this documentation section.