Adding a link to Plutus resources on Cardano Docs (#809)
Hello :) We have recently created a new section on Cardano Docs, which serves as a Plutus index - a list of different references to Plutus materials. We will be growing this list & also asking the community to contribute.
I have just added a link to it from your Plutus overview. Thanks so much