Update register-stake-pool.md - removed * in if for UTXO script
Update register-stake-pool.md - removed * in if for UTXO script

I had * for my environment, where * is necessary, because after redirecting output from Docker to local files, it seems to add a /r at the end of the fullUtxo.out file. This is not necessary here, as most users will not be using Docker.