node and cli documentation
This is part of an effort to put node and cli documentation on a single point. The community finds developers.cardano.org as the most trusted place for finding end user documentaiton. Thus the node and cli dev teams would like to have developers.cardano.org as the source of truth for latest and most up-to-date documentaion.