Use `IOHKNIX_VERSION` to get correct blst version
Updated the command to fetch the correct `blst` version by parameterizing the
iohk-nix version with the `$IOHKNIX_VERSION` variable.

The documentation states that

> The variable `IOHKNIX_VERSION` is used to retrieve the correct versions of `sodium`, `secp256k1` and `blst`.

However this is not currently the case for `blst` version, which is retrieved from
iohk-nix master, and therefore the version can be incorrect for the given node
revision.