fix(docs): correct install commands for Fedora and GHCup
Update Fedora package list to use 'gmp-devel' instead of 'gmp-fdevel'
(typo).

Fix GHCup install command syntax for GHC and Cabal. With the incorrect
placement of the `--set` flag, the installation fails with:

```
[ Warn  ] This is an old-style command for installing GHC. Use 'ghcup install ghc' instead.
[ Error ] [GHCup-07140] Both installation and setting the tool failed.
[ ...   ] Install error was: Unable to find a download for GHC version 'ghc' on detected platform x86_64-linux-ubuntu-22.04
[ ...   ] Set error was: The version 'ghc' of the tool ghc is not installed.
[ Error ] Also check the logs in /root/.ghcup/logs
```

Update cabal build command for cardano-cli to match correct target.
Without this change, the build fails with:

```
Error: [Cabal-7130]
Internal error in target matching: could not make an unambiguous fully qualified target selector for 'exe:cardano-cli'.
We made the target 'exe:cardano-cli' (unknown-component) that was expected to be unambiguous but matches the following targets:
'exe:cardano-cli', matching:
  - exe:cardano-cli (unknown-component)
  - :pkg:exe:lib:exe:file:cardano-cli (unknown-file)

Note: Cabal expects to be able to make a single fully qualified name for a target or provide a more specific error. Our failure to do so is a bug in cabal. Tracking issue: https://github.com/haskell/cabal/issues/8684

Hint: this may be caused by trying to build a package that exists in the project directory but is missing from the 'packages' stanza in your cabal project file.

Error: cardano-cli not built
```